+++ CHANGELOG.md
@@ -4,6 +4,7 @@ Yii Framework 2 bootstrap4 extension Change Log
 2.0.2 under development
 -----------------------
 
+- Bug #121: Fixed Progress widget with e.g. russian locale (simialbi)
 - Enh #120: Added Pagination bs4 (simialbi)